+++ backend/package.json
@@ -4,7 +4,9 @@
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node server.js",
+    "startNodemon": "nodemon server.js"
+
   },
   "keywords": [],
   "author": "",
